+++ anymotion_cli/core.py
@@ -8,7 +8,7 @@ from . import __version__
 from .commands.analysis import cli as analysis
 from .commands.analyze import cli as analyze
 from .commands.compare import cli as compare
-# from .commands.comparison import cli as comparison
+from .commands.comparison import cli as comparison
 from .commands.configure import cli as configure
 from .commands.download import cli as download
 from .commands.draw import cli as draw
@@ -36,7 +36,7 @@ class ColorsCommandCollection(HelpColorsMixin, click.CommandCollection):
         analysis,
         analyze,
         compare,
-        # comparison,
+        comparison,
         configure,
         download,
         draw,